requirements.txt
Django == 2.2.12
djangorestframework == 3.13.1
google_api_python_client == 2.51.0
httplib2 == 0.14.0
oauth2client == 4.1.3
psycopg2 == 2.8.6
pyTelegramBotAPI == 4.4.0
requests == 2.22.0
xmltodict == 0.13.0
